Reimplement Qt5 ColorOverlay using Qt6 MultiEffect
https://stackoverflow.com/questions/70809206/color-overlay-an-image-in-qml-qt6#77060406
Even though the answer is not checked, but the Qt6.8 works, and after
following the suggestion in its comment about brightness.

The MultiEffect needs to be put outside of the Image when they're on a
Button. Having MultiEffect as the Image layer's effect only works when
it's not on a Button.